# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.TotalCostResponse:
    properties:
      period:
        type: string
      service:
        type: string
      total_cost:
        type: integer
      user_id:
        type: string
    type: object
  models.UpdateCreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST сервис для управления подписками пользователей
  title: Subscription API Service
  version: 1.0.0
paths:
  /subscriptions:
    get:
      description: Возвращает список всех подписок с пагинацией
      parameters:
      - description: Количество записей (по умолчанию 10)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка в параметрах
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список подписок
      tags:
      - Subscriptions
    post:
      consumes:
      - application/json
      description: Создает запись о новой подписке пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Подписка успешно создана
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новую подписку
      tags:
      - Subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Подписка удалена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку
      tags:
      - Subscriptions
    get:
      description: Возвращает информацию о конкретной подписке
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Подписка найдена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - Subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет данные подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Подписка обновлена
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Подписка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить существующую подписку
      tags:
      - Subscriptions
  /subscriptions/total-cost:
    get:
      description: Возвращает суммарную стоимость всех подписок за указанный период
        с возможностью фильтрации
      parameters:
      - description: Начало периода (MM-YYYY)
        in: query
        name: start_date
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: end_date
        required: true
        type: string
      - description: Фильтр по ID пользователя
        in: query
        name: user_id
        type: string
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Суммарная стоимость
          schema:
            $ref: '#/definitions/models.TotalCostResponse'
        "400":
          description: Ошибка в параметрах
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Рассчитать стоимость подписок за период
      tags:
      - Subscriptions
schemes:
- http
swagger: "2.0"
tags:
- description: Операции с подписками пользователей
  name: Subscriptions